+++ core_client/base/models/v3/__init__.py
@@ -112,6 +112,10 @@ from .process_config import ProcessConfig
 from .process_probe_stream import ProcessProbeStream
 from .process_probe import ProcessProbe
 
+from .playout_status_io import PlayoutStatusIO
+from .playout_status_swap import PlayoutStatusSwap
+from .playout_status import PlayoutStatus
+
 from .process_state_exec import ProcessStateExec
 from .process_state_order import ProcessStateOrder
 from .process_state_progress_io_avstream_io_state import (