+++ core_client/base/models/v3/playout_status_swap.py
@@ -1,0 +1,17 @@
+from pydantic import BaseModel
+
+
+class PlayoutStatusSwap(BaseModel):
+    """
+    {
+        "lasterror": "string",
+        "lasturl": "string",
+        "status": "string",
+        "url": "string"
+    }
+    """
+
+    lasterror: str | None = None
+    lasturl: str | None = None
+    status: str | None = None
+    url: str | None = None